% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlus.R
\name{plot_volcano_plus}
\alias{plot_volcano_plus}
\title{Create a custom volcano plot with asymptotic thresholds and optional labeling}
\usage{
plot_volcano_plus(
  to_plot,
  title,
  significance_thresholds = list(horizontal_asymptote = 2, vertical_asymptote = 0.5),
  labeling_criteria = NULL
)
}
\arguments{
\item{to_plot}{A data frame containing at least the following columns:
\code{logFC} (log fold change) and \code{adj.P.Val} (adjusted p-value).
Optionally, a \code{Gene} column can be included for labeling specific genes.}

\item{title}{The title of the volcano plot.}

\item{significance_thresholds}{A named list of parameters for defining the significance threshold function.
This list should contain:
\itemize{
\item \code{horizontal_asymptote}: Controls the horizontal threshold for significance (default is \code{2}).
\item \code{vertical_asymptote}: Controls the vertical threshold for significance (default is \code{0.5}).
}}

\item{labeling_criteria}{An optional named list of parameters for labeling specific points.
This list should contain:
\itemize{
\item \code{horizontal_asymptote}: Controls the horizontal threshold for labeling (default is \code{1.5}).
\item \code{vertical_asymptote}: Controls the vertical threshold for labeling (default is \code{0.4}).
If not provided, no points will be labeled.
}}
}
\value{
A \code{ggplot} object representing the volcano plot.
}
\description{
\code{plot_volcano_plus} generates a volcano plot using \code{ggplot2}, applying custom
asymptotic thresholds for identifying significant points, and optionally labeling
specific points based on additional criteria.
}
\details{
The function identifies significant points by applying a mirrored
asymptotic function to the log fold change (\code{logFC}) values. Points that fall below
this threshold are considered insignificant and are plotted with reduced transparency.
Optionally, specific points can be labeled based on additional asymptotic criteria.
}
\examples{
# Example usage:
# Assuming `data` is a data frame with `logFC`, `adj.P.Val`, and `Gene` columns.
significance_thresholds <- list("horizontal_asymptote" = 2, "vertical_asymptote" = 0.5)
labeling_criteria <- list("horizontal_asymptote" = 1.5, "vertical_asymptote" = 0.4)
plot_volcano_plus(data, "Volcano Plot Example", significance_thresholds, labeling_criteria)
}
